extracted getFormData
